upgrade rama to 0.3 (requires MSRV rust 1.96)
