--- benchmark/web/rust/rama-0.2-rust-1.86/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "app"
-version = "0.1.0"
-edition = "2024"
-publish = false
-rust-version = "1.86.0"
-
-[features]
-default = []
-tracing = ["dep:tracing", "dep:tracing-subscriber"]
-
-[dependencies]
-rama = { version = "0.2.0", features = ["http-full", "compression"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tokio = { version = "1.0", features = ["full"] }
-tracing = { version = "0.1", optional = true }
-tracing-subscriber = { version = "0.3", features = [
-    "env-filter",
-], optional = true }
