+++ benchmark/web/rust/rama-0.3-rust-1.96/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "app"
+version = "0.1.0"
+edition = "2024"
+publish = false
+rust-version = "1.96.0"
+
+[features]
+default = []
+tracing = []
+
+[dependencies]
+rama = { version = "0.3.0", features = ["http-full", "compression"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tokio = { version = "1.0", features = ["full"] }
