+++ benchmark/web/rust/rama-0.3-rust-1.96/benchmark.yaml
@@ -1,9 +1,9 @@
 language: Rust
 mode: Default
 version:
-  - "1.86"
+  - "1.96"
 framework: Rama
 framework_website: https://github.com/plabayo/rama
 framework_flavor: Default
 framework_version:
-  - "0.2.0"
+  - "0.3.0"
